# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RaktOra REST API",
    "description": "Comprehensive REST API specification for RaktOra - National Voluntary Blood Network and Smart FEFO Inventory Engine.",
    "version": "1.0.0",
    "contact": {
      "name": "RaktOra Developer Team",
      "email": "support@raktora.org"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base API Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token"
      }
    },
    "schemas": {
...
      "AppointmentInput": {
        "type": "object",
        "required": ["blood_bank_id", "appointment_date", "time_slot"],
        "properties": {
          "blood_bank_id": { "type": "integer", "example": 1 },
          "appointment_date": { "type": "string", "format": "date", "example": "2026-09-01" },
          "time_slot": { "type": "string", "example": "10:00 AM - 11:00 AM" },
          "donation_type": { "type": "string", "enum": ["WHOLE_BLOOD", "PRBC", "PLATELETS", "FFP", "CRYOPRECIPITATE", "SDP"], "example": "WHOLE_BLOOD" }
        }
      },
...
    }
  },
  "paths": {
...
    "/appointments": {
      "get": {
        "summary": "List Appointments",
        "tags": ["Appointments"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Appointments list" }
        }
      },
      "post": {
        "summary": "Book Blood Donation Appointment",
        "tags": ["Appointments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AppointmentInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Appointment booked successfully" }
        }
      }
    },
...
  }
}